apiadmin/requirements.txt
django==2.2.3
django-appconf==1.0.3
django-crispy-forms==1.7.2
django-image-cropping==1.2.0
djangorestframework==3.10.2
easy-thumbnails==2.6
html5lib==1.0.1
Pillow==6.1.0
PyPDF2==1.26.0
pytz==2019.1
reportlab==3.5.23
six==1.12.0
sqlparse==0.3.0
webencodings==0.5.1
xhtml2pdf==0.2.3
